Add optional result backend, add TaskGraphs, rework callback publishing (#19)
This PR includes a large number of changes to this project, described in greater detail below. In brief, the changes here include:

* Refactor of task-handling logic
* Add optional result-storage for storing task results in Azure Blob Storage
* Add TaskGraphs for evaluating DAGs (relies on above result-storage)

This PR is marked as a development release with version v1.0.0-dev1. This feature is not intended to be published for users as-is.

## New Evaluator classes

- "simple"
- Results Storage
- TaskGraph

These classes are used to process *each* task based on whether these tasks are members of a graph or whether a StorageInterface has been provided to Boilermaker. 

**"Simple" represents the status quo *before* this PR is merged: and existing functionality should be preserved here without any changes required by callers.**

## Future Work

In the future, we can offer a few different improvements on this work:

* TaskGraph visualization (graphviz, mermaid, web-renderings, etc.) including completion percentages for "live" 
* Greater ergonomics around assembling TaskGraphs: it's not easy at the moment to use this feature and requires in-depth knowledge of the data-structures in this codebase.
* Automatic results-storage cleanup (async timer, script, etc.)
* Allow passing the return value of one task to its children as inputs.
* Allow parallel or batch processing in a single worker.